+++ src/Infrastructure/Elastic/ElasticIndexAdapter.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JeroenG\Explorer\Infrastructure\Elastic;
+
+use Elasticsearch\Client;
+use JeroenG\Explorer\Application\IndexAdapterInterface;
+use JeroenG\Explorer\Domain\IndexManagement\IndexConfigurationInterface;
+
+final class ElasticIndexAdapter implements IndexAdapterInterface
+{
+    private Client $client;
+
+    public function __construct(ElasticClientFactory $clientFactory)
+    {
+        $this->client = $clientFactory->client();
+    }
+
+    public function create(IndexConfigurationInterface $indexConfiguration): void
+    {
+        $this->client->indices()->create($indexConfiguration->toArray());
+    }
+
+    public function delete(IndexConfigurationInterface $indexConfiguration): void
+    {
+        $this->client->indices()->delete(['index' => $indexConfiguration->getName()]);
+    }
+
+    public function flush(string $index): void
+    {
+        $matchAllQuery = [ 'query' => [ 'match_all' => (object)[] ] ];
+        $this->client->deleteByQuery([
+            'index' => $index,
+            'body' => $matchAllQuery
+        ]);
+    }
+}